% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{ZarrGroup}
\alias{ZarrGroup}
\title{A handle to a Zarr group (root or sub-group) within a store.}
\format{
An object of class \code{Rzarrs::ZarrGroup__bundle} (inherits from \code{savvy_Rzarrs__sealed}) of length 2.
}
\usage{
ZarrGroup
}
\description{
Groups may contain arrays and/or sub-groups.  Use `$attributes()` to read
the group's JSON attributes as a native R list, and `$children()` to list
the immediate child nodes.
}
\keyword{datasets}
