% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\docType{data}
\name{ZarrVcf}
\alias{ZarrVcf}
\title{High-level VCF Zarr reader}
\format{
An object of class \code{environment} of length 1.
}
\usage{
ZarrVcf
}
\arguments{
\item{x}{Path (string), URL (string), `ZarrStore`, or `ZarrObjectStore`.}
}
\value{
A `ZarrVcf` instance.
}
\description{
`ZarrVcf$open(x)` opens a VCF Zarr store and returns an instance with
methods to access variant, sample, and genotype data. Supports spec
versions 0.1 through 0.4.
}
\keyword{datasets}
